+++ scripts/macros/2024/mechanics/invisible.js
@@ -3,14 +3,12 @@ async function attacked({trigger, workflow}) {
     if (!workflow.targets.size || !constants.attacks.includes(workflow.activity.actionType) || !workflow.token) return;
     if (tokenUtils.canSense(workflow.targets.first(), workflow.token, ['feelTremor', 'seeInvisbility', 'blindsight', 'seeAll', 'senseAll', 'senseInvisibility'])) return;
     workflow.disadvantage = true;
-    workflow.rollOptions.disadvantage = true;
     workflow.attackAdvAttribution.add('DIS: ' + trigger.entity.name);
 }
 async function attacking({trigger, workflow}) {
     if (!workflow.targets.size || !constants.attacks.includes(workflow.activity.actionType) || !workflow.token) return;
     if (tokenUtils.canSense(workflow.targets.first(), workflow.token, ['feelTremor', 'seeInvisbility', 'blindsight', 'seeAll', 'senseAll', 'senseInvisibility'])) return;
     workflow.advantage = true;
-    workflow.rollOptions.advantage = true;
     workflow.attackAdvAttribution.add('ADV: ' + trigger.entity.name);
 }
 export let invisible = {